add logining

+++ updater.go
@@ -14,16 +14,46 @@ import (
 	"os/exec"
 	"runtime"
 	"strings"
+	"sync"
 	"time"
 )
 
 const configFile = "conf.txt"
-const key = "2112751343910010"
+const key = "2112751343910000"
+const logFileName = "log.txt"
 
+var clientExist = false
 var conf config
 var clientApp = "client"
 var newClientApp = "newClient"
 var command = "./" + clientApp
+
+var l = logData{fileName: logFileName}
+
+type logData struct {
+	m        sync.Mutex
+	fileName string
+	eol      string
+}
+
+func toLog(data string, flag bool) {
+	data = "updater " + time.Now().Format("02.01.2006 15:04:05") + " " + data
+	l.m.Lock()
+	file, err := os.OpenFile(l.fileName, os.O_APPEND|os.O_WRONLY, 0666)
+	if err != nil {
+		log.Fatal("open log file")
+	}
+	_, err = file.WriteString(data + l.eol)
+	if err != nil {
+		file.Close()
+		log.Fatal("write data to log")
+	}
+	file.Close()
+	l.m.Unlock()
+	if flag {
+		log.Fatal(data)
+	}
+}
 
 type config struct {
 	UpdaterServer string
@@ -41,18 +71,29 @@ type clientData struct {
 	System   string
 }
 
-func check(err error) {
-	if err != nil {
-		log.Fatal(err)
-	}
-}
-
 func init() {
+	if !ex.ExistFile(l.fileName) {
+		file, err := os.OpenFile(l.fileName, os.O_CREATE, 0666)
+		if err != nil {
+			log.Fatal("create log file")
+		}
+		file.Close()
+	}
+	if runtime.GOOS == "windows" {
+		l.eol = "\r\n"
+	}
+	if runtime.GOOS == "linux" {
+		l.eol = "\n"
+	}
 	if ex.ExistFile(configFile) {
 		data, err := ex.ReadFileFull(configFile)
-		check(err)
+		if err != nil {
+			toLog("init: read conf file", true)
+		}
 		err = json.Unmarshal(data, &conf)
-		check(err)
+		if err != nil {
+			toLog("init: unmarshal conf file", true)
+		}
 	} else {
 		conf := config{
 			UpdaterServer: "127.0.0.1:50000",
@@ -62,16 +103,22 @@ func init() {
 			ClientId:      "----------------",
 		}
 		data, err := json.MarshalIndent(&conf, "", "  ")
-		check(err)
+		if err != nil {
+			toLog("init: marshal conf file", true)
+		}
 		err = ioutil.WriteFile(configFile, data, 0644)
-		check(err)
+		if err != nil {
+			toLog("init: write conf file", true)
+		}
 		log.Fatal("Файл конфигурации не найден. Создан новый файл конфигурации.")
 	}
 	if runtime.GOOS == "windows" {
 		clientApp += ".exe"
 		newClientApp += ".exe"
 		path, err := os.Executable()
-		check(err)
+		if err != nil {
+			toLog("init: os.Executable()", true)
+		}
 		i := strings.LastIndex(path, "\\")
 		path = strings.Replace(path, path[i+1:], "", 1)
 		clientApp = path + clientApp
@@ -80,7 +127,10 @@ func init() {
 	}
 	path, err := os.Stat(clientApp)
 	if err != nil || path.IsDir() {
-		log.Fatal("no client")
+		fmt.Println("no client")
+		conf.VersionClient = "-----"
+	} else {
+		clientExist = true
 	}
 }
 
@@ -100,9 +150,8 @@ func newClient() *clientData {
 
 //Получает файл актуального клиента
 func (cl *clientData) downloadNewClient(q cn.Query) error {
-	var err error
 	cn.SendSync(cl.conn)
-	err = cn.GetFile(q.Query, q.DataLen, cl.conn)
+	err := cn.GetFile(q.Query, q.DataLen, cl.conn)
 	if err != nil {
 		return errors.New("downloadNewClient")
 	}
@@ -112,19 +161,27 @@ func (cl *clientData) downloadNewClient(q cn.Query) error {
 //Подключается к серверу и получаетот него указание
 func (cl *clientData) connect() error {
 	if !cl.validOnServer(cl.conn) {
-		log.Fatal("Valid on Server")
+		toLog("connect: Valid on Server", true)
 	}
 	err := cn.SendString(conf.ClientId, cl.conn)
+	if err != nil {
+		return errors.New("connect: SendString(conf.ClientId, cl.conn)")
+	}
 	cn.ReadSync(cl.conn)
 	jsonData, err := json.Marshal(cl)
+	if err != nil {
+		return errors.New("connect: json.Marshal(cl)")
+	}
 	err = cn.SendBytesWithDelim(jsonData, cl.conn)
 	if err != nil {
-		return err
+		return errors.New("connect: SendBytesWithDelim(jsonData, cl.conn)")
 	}
 	q, err := cn.ReadQuery(cl.conn)
 	if err != nil {
-		return err
+		return errors.New("connect: ReadQuery(cl.conn)")
 	}
+	toLog(fmt.Sprintf("Query: %#v", q), false)
+	fmt.Printf("Query: %#v\n", q)
 	switch q.Method {
 	case "already exist":
 		log.Fatal("already exist")
@@ -135,28 +192,56 @@ func (cl *clientData) connect() error {
 		}
 		path, err := os.Stat(newClientApp)
 		if err == nil && !path.IsDir() {
+			toLog("update client", false)
 			fmt.Println("update client")
-			err = os.Remove(clientApp)
-			check(err)
+			if clientExist {
+				err = os.Remove(clientApp)
+				if err != nil {
+					return errors.New("connect: os.Remove(clientApp)")
+				}
+			}
+			time.Sleep(time.Second)
 			err := os.Rename(newClientApp, clientApp)
-			check(err)
+			if err != nil {
+				return errors.New("connect: os.Rename(newClientApp, clientApp)")
+			}
+			time.Sleep(time.Second)
 			if runtime.GOOS != "windows" {
-				time.Sleep(time.Second)
 				err = os.Chmod(clientApp, 0777)
-				check(err)
+				if err != nil {
+					return errors.New("connect: os.Chmod(clientApp, 0777)")
+				}
 				time.Sleep(time.Second)
 			}
 		} else {
-			log.Fatal("newClientApp: if err == nil && !path.IsDir()")
+			toLog("connect: newClientApp: if err == nil && !path.IsDir()", true)
 		}
 		return nil
 	case "lenClient":
 		data, err := ex.ReadFileFull(clientApp)
-		check(err)
+		if err != nil {
+			toLog("connect: ReadFileFull(clientApp)", true)
+		}
 		if len(data) != q.DataLen {
-			return errors.New("wrong client")
+			return errors.New("connect: wrong client")
 		}
 		return nil
+	case "getLog":
+		data, err := ex.ReadFileFull(logFileName)
+		if err != nil {
+			err = cn.SendQuery(cn.Query{Query: "err read log file"}, cl.conn)
+			toLog("connect: read log file Full", true)
+		}
+		err = cn.SendQuery(cn.Query{DataLen: len(data)}, cl.conn)
+		if err != nil {
+			return errors.New("connect: cn.SendQuery(cn.Query{DataLen: len(fileBytes)}, cl.conn)")
+		}
+		cn.ReadSync(cl.conn)
+		err = cn.SendBytes(data, cl.conn)
+		if err != nil {
+			return errors.New("connect: cn.SendBytes(data, cl.conn)")
+		}
+		return errors.New("send logFile")
 	}
 	return nil
 }
@@ -165,12 +250,17 @@ func (cl *clientData) connect() error {
 func (cl *clientData) validOnServer(conn net.Conn) bool {
 	var code = make([]byte, 16)
 	bc, err := aes.NewCipher([]byte(key))
+	if err != nil {
+		toLog("validOnServer: aes.NewCipher([]byte(key))", true)
+	}
 	err = cn.SendString("updater", conn)
 	if err != nil {
+		toLog("validOnServer: SendString(\"updater\", conn)", false)
 		return false
 	}
 	data, err := cn.ReadBytesByLen(16, conn)
 	if err != nil {
+		toLog("validOnServer: ReadBytesByLen(16, conn)", false)
 		return false
 	}
 	bc.Decrypt(code, data)
@@ -179,10 +269,15 @@ func (cl *clientData) validOnServer(conn net.Conn) bool {
 	bc.Encrypt(code, []byte(res))
 	err = cn.SendBytes(code, conn)
 	if err != nil {
+		toLog("validOnServer: SendBytes(code, conn)", false)
 		return false
 	}
 	mes, err := cn.ReadString(conn)
-	if err != nil || mes != "ok" {
+	if err != nil {
+		toLog("validOnServer: ReadString(conn)//res of valid", false)
+		return false
+	}
+	if mes != "ok" {
 		return false
 	}
 	return true
@@ -203,28 +298,35 @@ func clientRun() int {
 
 func main() {
 	cl := newClient()
+	toLog("Start updater", false)
 	fmt.Println("Start updater")
 	for {
 		conn, err := net.Dial("tcp", conf.UpdaterServer)
 		if err != nil {
-			fmt.Println("Server not found")
-			time.Sleep(5 * time.Second)
+			for i := 5; i >= 0; i-- {
+				fmt.Printf("Server not found. Time to reconnect: %d\r", i)
+				time.Sleep(1 * time.Second)
+			}
 			continue
 		}
+		toLog(fmt.Sprint("connect to conf.UpdaterServer: ", conf.UpdaterServer), false)
 		cl.conn = conn
 		err = cl.connect()
 		if err != nil {
+			toLog(fmt.Sprint(err), false)
 			fmt.Println(err, "sleep min.")
 			cl.conn.Close()
 			time.Sleep(time.Minute)
 			continue
 		}
 		if exitCode := clientRun(); exitCode > 0 {
+			toLog("exitCode. sleep min.", false)
 			fmt.Println("exitCode. sleep min.")
 			cl.conn.Close()
 			time.Sleep(time.Minute)
 			continue
 		}
+		toLog("No exitCode. sleep min.", false)
 		fmt.Println("No exitCode. sleep min.")
 		cl.conn.Close()
 		time.Sleep(time.Minute)
